various unit tests and bug fixes. Compiles and runs, but there is a bug with pairHMM likelihood calc (reads too long?), some samples break graph construction (Losing reference source and sink nodes, or having too many), and AssemblyResultSet needs to have better handles for multiple kmer sizes

+++ src/utils/artificial_read_utils.rs
@@ -68,7 +68,7 @@ impl ArtificialReadUtils {
     pub fn create_artificial_read_default(
         name: &str,
         tid: usize,
-        alignment_start: usize,
+        alignment_start: i64,
         length: usize,
         unmapped: bool,
     ) -> Record {
@@ -85,7 +85,7 @@ impl ArtificialReadUtils {
             seq.as_slice(),
             quals.as_slice(),
         );
-        record.set_pos(alignment_start as i64);
+        record.set_pos(alignment_start);
         record.set_tid(tid as i32);
         record.unset_proper_pair();
